# Conflicts:
#	Tools/QuickEd/CMakeLists.txt
#	Tools/QuickEd/Data/Materials/Shaders/2d/composit-fp.cg
#	Tools/QuickEd/Data/Materials/Shaders/Default/grass-vp.cg
#	Tools/QuickEd/Data/Materials/Shaders/Default/materials-fp.cg
#	Tools/QuickEd/Data/Materials/Shaders/Default/materials-vp.cg
#	Tools/QuickEd/Data/Materials/Shaders/Default/water-fp.cg
#	Tools/QuickEd/Data/Materials/Shaders/Default/water-vp.cg
#	Tools/QuickEd/Data/Materials/Shaders/Landscape/landscape-editor-vp.cg
#	Tools/QuickEd/Data/Materials/Shaders/Landscape/tilemask-fp.cg
#	Tools/QuickEd/Data/Materials/Shaders/Landscape/tilemask-vp.cg
#	Tools/QuickEd/Data/Materials/Shaders/common.cgh
#	Tools/QuickEd/Data/Materials/assignable.yaml
#	Tools/ResourceEditor/CMakeLists.txt
#	Tools/ResourceEditor/Classes/Qt/Scene/System/CustomColorsSystem.cpp
#	Tools/ResourceEditor/Classes/Qt/Scene/System/HeightmapEditorSystem.cpp
#	Tools/ResourceEditor/Classes/Qt/Scene/System/LandscapeEditorDrawSystem/LandscapeProxy.cpp
#	Tools/ResourceEditor/Classes/Qt/Scene/System/RulerToolSystem.cpp
#	Tools/ResourceEditor/Classes/Qt/Scene/System/TilemaskEditorSystem.cpp
#	Tools/ResourceEditor/Classes/Qt/main.cpp


Former-commit-id: 5ad6018d0385ce868f9c5b281e292164c81ad332

+++ Tools/Launcher/Classes/filemanager.cpp
@@ -1,4 +1,5 @@
 #include "filemanager.h"
+#include "errormessenger.h"
 #include <QDesktopServices>
 #include <QCoreApplication>
 #include <QDir>
@@ -104,6 +105,11 @@ bool CreateFileAndWriteData(const QString& filePath, const QByteArray& data)
 
 bool DeleteDirectory(const QString& path)
 {
+    if (path == "/" || path == "." || path == "..")
+    {
+        ErrorMessenger::ShowErrorMessage(ErrorMessenger::ERROR_PATH, "trying to remove wrong path! Aborted");
+        return false;
+    }
     QDir dir(path);
     return dir.removeRecursively();
 }
@@ -250,18 +256,15 @@ void MakeDirectory(const QString& path)
 QString GetApplicationDirectory(const QString& branchID, const QString& appID)
 {
     QString path = GetBranchDirectory(branchID) + appID + "/";
-    if (!QDir(path).exists())
-        QDir().mkpath(path);
-
     return path;
 }
 
 QString GetBranchDirectory(const QString& branchID)
 {
-    QString path = GetBaseAppsDirectory() + branchID + "/";
-    if (!QDir(path).exists())
-        QDir().mkpath(path);
-
+    QString dirName = branchID;
+    dirName.remove("/");
+    dirName.remove("\\");
+    QString path = GetBaseAppsDirectory() + dirName + "/";
     return path;
 }
 }
